feat(@covalent/coding-standards): rename to @covalent/coding-standards
