+++ .huskyrc.js
@@ -1,2 +1,2 @@
-var defaultConfig = require('./src/platform/standards/husky/husky.js');
+var defaultConfig = require('./src/platform/coding-standards/husky/husky.js');
 module.exports = defaultConfig;
