+++ commitlint.config.js
@@ -1,2 +1,2 @@
-var defaultConfig = require('./src/platform/standards/commitlint/commitlint.config.js');
+var defaultConfig = require('./src/platform/coding-standards/commitlint/commitlint.config.js');
 module.exports = defaultConfig;
