+++ package.json
@@ -99,7 +99,7 @@
     "@angular-devkit/build-ng-packagr": "~0.801.2",
     "@angular/cli": "8.1.2",
     "@angular/compiler-cli": "^8.1.3",
-    "@covalent/standards": "file:src/platform/standards",
+    "@covalent/coding-standards": "file:src/platform/coding-standards",
     "@types/fs-extra": "^4.0.0",
     "@types/hammerjs": "^2.0.30",
     "@types/jasmine": "~2.8.6",
