+++ prettier.config.js
@@ -1,2 +1,2 @@
-var defaultConfig = require('./src/platform/standards/prettier/prettier.config.js');
+var defaultConfig = require('./src/platform/coding-standards/prettier/prettier.config.js');
 module.exports = defaultConfig;
