+++ scripts/post-ng-packagr-build.js
@@ -16,8 +16,10 @@ gulp.task('move-additional-platform-files', function() {
   return gulp.src(config.paths.PostNgPackngrAdditionalFiles).pipe(gulp.dest(config.deployed));
 });
 
-gulp.task('move-standards-package', function() {
-  return gulp.src(config.paths.PostNgPackagerStandardsPackage).pipe(gulp.dest(`${config.deployed}standards`));
+gulp.task('move-coding-standards-package', function() {
+  return gulp
+    .src(config.paths.PostNgPackagerCodingStandardsPackage)
+    .pipe(gulp.dest(`${config.deployed}coding-standards`));
 });
 
 gulp.task('compile-core-sass', function() {
