+++ src/platform/coding-standards/husky/husky.js
