+++ src/platform/coding-standards/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@covalent/standards",
+  "name": "@covalent/coding-standards",
   "version": "0.0.0-COVALENT",
   "description": "Teradata Covalent shared configurations",
   "keywords": [
