+++ src/platform/coding-standards/prettier/README.md
@@ -24,7 +24,7 @@ Update appropriate linter configs to have the newly installed linter rule sets a
 
 ```json
 {
-  "extends": ["./node_modules/@covalent/standards/tslint/tslint.json", "tslint-config-prettier"]
+  "extends": ["./node_modules/@covalent/coding-standards/tslint/tslint.json", "tslint-config-prettier"]
 }
 ```
 
@@ -32,7 +32,7 @@ Update appropriate linter configs to have the newly installed linter rule sets a
 
 ```json
 {
-  "extends": ["./node_modules/@covalent/standards/stylelint/stylelint.config.js", "stylelint-config-prettier"]
+  "extends": ["./node_modules/@covalent/coding-standards/stylelint/stylelint.config.js", "stylelint-config-prettier"]
 }
 ```
 
@@ -41,7 +41,7 @@ Update appropriate linter configs to have the newly installed linter rule sets a
 Create a `prettier.config.js` file at the root level that references the `prettier.config.js` file located inside this repo.
 
 ```javascript
-var defaultConfig = require('./node_modules/@covalent/standards/prettier/prettier.config.js');
+var defaultConfig = require('./node_modules/@covalent/coding-standards/prettier/prettier.config.js');
 module.exports = defaultConfig;
 ```
