+++ src/platform/coding-standards/prettier/prettier.config.js
