+++ src/platform/coding-standards/stylelint/selector-no-deprecated-deep/index.js
