+++ src/platform/coding-standards/stylelint/stylelint.config.js
