+++ src/platform/coding-standards/tslint/tslint.json
