+++ stylelint.config.js
@@ -1,4 +1,4 @@
 module.exports = {
-  extends: ['./src/platform/standards/stylelint/stylelint.config.js'],
+  extends: ['./src/platform/coding-standards/stylelint/stylelint.config.js'],
   ignoreFiles: './src/platform/core/schematics/ng-add/files/**',
 };
