+++ tslint.json
@@ -1,3 +1,3 @@
 {
-  "extends": ["./src/platform/standards/tslint/tslint.json"]
+  "extends": ["./src/platform/coding-standards/tslint/tslint.json"]
 }
